Move the `IncludedBuildTaskResource` implementations closer to the work graph for the build that owns and executes the task.
